backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
async-timeout==4.0.3
attrs==25.4.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
build==1.4.0
cachetools==5.5.2
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
dataclasses-json==0.6.7
distro==1.9.0
durationpy==0.10
exceptiongroup==1.3.1
faiss-cpu==1.8.0
filelock==3.24.3
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
googleapis-common-protos==1.72.0
greenlet==3.3.2
grpcio==1.78.1
h11==0.16.0
hf-xet==1.3.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==0.36.2
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==0.2.7
langchain-community==0.2.5
langchain-core==0.2.43
langchain-text-splitters==0.2.4
langsmith==0.1.147
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
narwhals==2.17.0
networkx==3.4.2
numpy==1.26.4
orjson==3.11.7
overrides==7.7.0
packaging==24.2
pandas==2.2.2
pillow==10.4.0
pip==26.0.1
propcache==0.4.1
protobuf==5.29.6
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.8.0
pypdf==4.2.0
PyPDF2==3.0.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.19
requests==2.32.5
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==2.7.0
setuptools==69.1.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.47
streamlit==1.38.0
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
tomli==2.4.0
torch==2.0.1+cpu
tornado==6.5.4
tqdm==4.66.4
transformers==4.57.6
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==1.26.18
uvloop==0.22.1
watchdog==4.0.2
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
wheel==0.42.0
wrapt==1.17.3
yarl==1.22.0
zhipuai==2.1.5.20250825
zipp==3.23.0
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
pydantic-settings  
python-multipart   